series_sort: Catch pygit2 import failure.
Provide some instructions about how to install the module in the README.

Note that this only catches the exception from series_sort.py. There are
other executable python scripts which import pygit2 and those will still
raise an ImportError if the pygit2 module is not available. However,
series_sort.py is the most common entry point for new users.