disable patches.drivers/s390-qeth-use-Read-device-to-query-hypervisor-for-MA.patch
Backport of mainline commit b7493e91c11a ("s390/qeth: use Read device to
query hypervisor for MAC") changes assigned MAC address (and breaks
networking) on one of our machines and it's not clear which address is
actually correct (bsc#1094575). Disable the patch for now with a marker so
that we prevent releasing a maintenance update incompatible with GM. Once
the bug is resolved, we will either reenable the patch or drop it.