Update Patch-mainline tags to include rc information
Some Patch-mainline tags include the release candidate (rc) while others do
not. Because of that inconsistency it is impossible to know from a
Patch-mainline tag with no rc if it describes a commit that was added
between the last rc and the release of a development cycle or if the rc
part was not included in the tag. Update Patch-mainline tags of patches in
the sorted section to include rc information consistently.