Revert "move changes without Git-commit out of sorted section"
This reverts commit cbf049ebedc611cdffd443910686956c1f4b7554.

The log for cbf049ebedc61 is:
	move changes without Git-commit out of sorted section

However, the purpose of the out-of-tree section is precisely to contain
patches without Git-commit tags.

After discussion, it seems that this commit was made in an attempt to fix
an unrelated and unreproduced problem.

Patches which are in "Patch-mainline: Submitted" state are moved back to
the out-of-tree subsection while they wait for a Git-commit tag.
Patches which seem to be in "Patch-mainline: Never" state are moved to a
position relevant to their subsystem after the sorted section.