btrfs: move some btrfs upstream patches to sorted series section
For btrfs we have a lot of patches that are upstream but are in the
unsorted section of series.conf, for reasons I do not know. This change
moves some of them.

Such large amounts of btrfs upstream patches in our unsorted section is
making the work of backporting upstream fixes harder and harder, for
example I can't backport at least a handful of upstream fixes that landed
in 4.20-rcX if I add them to the sorted section. This will only encourage
people to keep adding backports from upstream to the unsorted section,
defeating the purpose of having a sorted section.

Some patches in the unsorted section, mostly related to qgroups, are
big and fat touching important files like fs/btrfs/inode.c and
fs/btrfs/extent-tree.c for example, which are updated very frequently
upstream. I really want to be able to add upstream fixes to the sorted
section, as not doing it will only make life harder for everyone else
too.

Tested with sequence-patch.sh that this new order succeeds.

Signed-off-by: Filipe Manana <fdmanana@suse.com>