- Refresh   patches.fixes/libata-ensure-host-is-free-d-on-error-exit-paths.patch. - Delete patches.fixes/libata-add-refcounting-to-ata_host.patch. - Delete patches.kabi/kabi-protect-struct-ata_host.patch.
The fix in 'patches.fixes/libata-add-refcounting-to-ata_host.patch.'
introduced a kref in the middle of 'struct ata_host' which the KABI patch
'patches.kabi/kabi-protect-struct-ata_host.patch' then moved to the end of
the structure in order to preserve the KABI for out of tree modules. But
this move to the end of the structure shadowed the dynamically allocated
'ports' array at the end of 'struct ata_host' and lead to kernel crashes
(bsc#1125109).